+++ Tests/LinuxMain.swift
@@ -961,7 +961,8 @@ extension ReporterTests {
         ("testCheckstyleReporter", testCheckstyleReporter),
         ("testJunitReporter", testJunitReporter),
         ("testHTMLReporter", testHTMLReporter),
-        ("testSonarQubeReporter", testSonarQubeReporter)
+        ("testSonarQubeReporter", testSonarQubeReporter),
+        ("testMarkdownReporter", testMarkdownReporter)
     ]
 }